requirements.txt
langchain==0.3.26
langchain-google-genai==2.1.8
sentence-transformers==4.1.0
pypdf==5.6.1
flask==3.1.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-community==0.3.26
- e.
